+++ src/workerBridge/bridge.ts
@@ -1,0 +1,7 @@
+import type workerBridge from './workerBridge'
+export const getSeguroInitData = (bridgeWorker: workerBridge | null) => {
+    if (bridgeWorker) {
+        return bridgeWorker.seguroInitData
+    }
+    return null
+}